requirements.txt
discord.py
requests
python-dotenv
pyVinted
pytz